Adding the commit-counter to the default versionstring.

+++ include/redux/version.hpp
@@ -20,14 +20,14 @@ namespace redux {
     /*!
      *   @fn         extern int getVersionNumber (void)
      *   @brief      Returns an integer representing the version
-     *   @details    The different version numbers are packed into an integer that can be used
-     *               to compare versions. E.g. v. 1.2.3 would be returned as the integer 66051
-     *               (= (1<<16)+(2<<8)+3)
+     *   @details    The different version numbers are packed into a 64-bit integer that can be used
+     *               to compare versions. E.g. v. 1.2.3-9 would be returned as the integer 4328718345
+     *               (= (1<<32)+(2<<24)+(3<<16)+9)
      *
      *   @author    Tomas Hillberg (hillberg@astro.su.se)
      *   @date      2013
      */
-    extern int getVersionNumber( void );
+    extern uint64_t getVersionNumber( void );
 
 
     /*!
@@ -38,7 +38,7 @@ namespace redux {
      *   @date      2013
      */
     extern std::string getVersionString( void );
-    extern std::string getVersionString( int );
+    extern std::string getVersionString( uint64_t );
 
 
     /*!
